requirements.txt
flexrest
flask-sqlalchemy
flask-testing
flask-login
flask-principal
flask
sqlalchemy-migrate
sqlalchemy-utils
sqlalchemy
pycrypto
formencode
bcrypt
pytz
itsdangerous
requests
python-dateutil
httplib2
ua-parser
user-agents
uwsgi
simplejson
